ascad Deletion for Authenticated Users #75
For a user to delete another  (admin --> entrant) we require a cloud function in firebase. This is paid

So instead, the accounts detail is delet. Events, waiting lists, and if the account were on any waiting lists.
The account is then disabled in the database stopping logins from that account.

Account still exists in both FireAuth and Database, but is disabled